[DBCP-562] avoids exposing password via JMX
